refactor: ensure all on-disk-format is prefixed with xal_odf_
To clearly signal, in the code, what is memory layout that needs to be
compatible with the XFS on-disk-format, then all of these are prefixed
with "xal_odf_".

The __be is dropped as it is redundant, by default, everything in
the on-disk-format is big-endian. Thus, instead, fields that are
little-endian are described with a comment rather than a typedef'ed
le32. With this, the __be* typedefs are removed.

Signed-off-by: Simon A. F. Lund <os@safl.dk>