refactor: drop index/flattening
The contiguous address space is unviable because the number of
inodes per allocation group is unknown and unevenly distributed. The
calculations fail since 'agblocks' represents the total blocks in an
allocation group, not those used for inodes. As a result, indexing
flattens the entire block address space instead of just the inode
address space.

Removing it again.

Signed-off-by: Simon A. F. Lund <os@safl.dk>