refactor(api,xal): expand pool-helpers, change inode-repr.
* Signature of process_dinode_extents() changed

Replaced the pool helper function which was typed-to 'xal_inode':

* xal_pool_claim()

with two explicitly typed helpers:

* xal_pool_claim_inodes()
* xal_pool_claim_extents()

An inode is either a directory containing child directories or files or
a file with extents. Instead of using separate counters and pointers,
they are combined into a union, as their usage is mutually exclusive.

Signed-off-by: Simon A. F. Lund <os@safl.dk>