feat: add helper xal_absbno_offset() computing offset given block
The filesystem block number is not sequential; it encodes an AG-relative
block number. Therefore, it cannot be directly multiplied by the block
size. Instead, the AG and AG-relative block number must be extracted
separately. This is similar to inode location encoding, though the 64
bits are used differently.

This function does so and returns the offset. The helper is especially
helpful when resolving e.g. btree-extents.

Signed-off-by: Simon A. F. Lund <os@safl.dk>