refactor: rename 'absbno' to 'fsbno'
The term "absolute block number" was originally chosen based on the
assumption that it referred to a raw, sequential block number. However,
after reviewing "XFS Algorithms & Data Structures," it is clear that
this entity corresponds to xfs_fsblock_t, which represents a filesystem
block number.

Accordingly, rename absbno to fsbno to align with the established XFS
nomenclature. The initially assumed representation would align more
closely with xfs_rfsblock_t, but that is not the type used here.

Signed-off-by: Simon A. F. Lund <os@safl.dk>