fix(setup_files): do not print status of 'dd'
This is done to reduce the amount of data-emitted. It is not an issue so far as there are not a lot of files.
However, to generate millions of files, then it starts to become taxing and also slows down the test-report.
Thus, do not print the output of dd, we do not care either, only if it fails.

Signed-off-by: Simon A. F. Lund <os@safl.dk>