fix(cli): do not print 'extents' for dirs
The internal inode representation carries content as either pointers
to other inodes or extent info. Naturally, when an inode represents a
directory, there are no extents; instead, there are pointers to other
inodes.
When an inode represents a file, the content is an array of extents.
The printer always printed extent info -- in other words, it treated the
list of pointers to children as extents, resulting in garbage output.
This fixes it.

Signed-off-by: Simon A. F. Lund <os@safl.dk>