feat(build): add option to disable debug logging
Using meson buildtype=debug currently also enables debug logging in xal
which is fairly verbose. This leads to xal becoming unusable when debug
symbols are enabled and xal is run on large inputs, as the large amount
of print statements slows down the process significantly.

Debug symbols (for gdb) and debug logging are two separate concerns
which should be able to be decoupled. Therefore, I add a meson option,
`debug-logging`, which can be set to false when compiling with
buildtype=debug to disable debug logging. As the default value for this
option is true, this commit doesn't change the way building xal works
now, but adds the option to remove logging if needed.

Example usage:
```
meson setup build --buildtype=debug -Ddebug-logging=false
```

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>