refactor(xal): move be specific fields to backend struct
A new backend field on the xal struct acts as a polymorphic field to
represent the fields necessary for the backend. For the xfs backend,
this includes the dinodes and the pointer to the xnvme device, and for
the fiemap backend, this includes the mountpoint string.

A base-struct, `xal_backend_base`, holds two function pointers that lets
the xal struct call index with the backend (meaning, building the
file-system tree from the root inode), and close the backend.

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>