feat(extents): add converters for xal_extents
This allows the users of the library to retrieve the extent information
in either bytes or blocks, using the LBA format on the device.
The returned xal_extent is a copy, meaning it does not modify the original
struct.

The return value of the converter functions are `const`, as a way to
prevent/discourage the user from using the returned xal_extent as
argument for the converter functions again, and getting completely
wrong results.

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>