feat(fiemap): add traversal implementation of xal_get_inode()
This is the implementation of getting inodes from a file path when xal
is opened with `opts.file_lookupmode == XAL_FILE_LOOKUPMODE_TRAVERSE`.
This is slightly more slow than the existing strategy, where a hash map
is used, but as it does not use extra memory, it is favorable as a
default.

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>